# Conflicts:
#	app/assets/javascripts/components/Icon.tsx
#	app/assets/javascripts/services/alertService.ts

+++ app/assets/javascripts/components/NoteView/NoteView.tsx
@@ -37,15 +37,10 @@ import { ActionsMenu } from '../ActionsMenu';
 import { HistoryMenu } from '../HistoryMenu';
 import { ComponentView } from '../ComponentView';
 import { PanelSide, PanelResizer, PanelResizeType } from '../PanelResizer';
+import { ElementIds } from '@/element_ids';
 
 const MINIMUM_STATUS_DURATION = 400;
 const TEXTAREA_DEBOUNCE = 100;
-
-const ElementIds = {
-  NoteTextEditor: 'note-text-editor',
-  NoteTitleEditor: 'note-title-editor',
-  EditorContent: 'editor-content',
-};
 
 type NoteStatus = {
   message?: string;
@@ -101,6 +96,7 @@ export const reloadFont = (monospaceFont?: boolean) => {
 type State = {
   availableStackComponents: SNComponent[];
   editorComponentViewer?: ComponentViewer;
+  editorComponentViewerDidAlreadyReload?: boolean;
   editorStateDidLoad: boolean;
   editorTitle: string;
   editorText: string;
@@ -445,17 +441,26 @@ export class NoteView extends PureComponent<Props, State> {
   }
 
   public editorComponentViewerRequestsReload = async (
-    viewer: ComponentViewer
+    viewer: ComponentViewer,
+    force?: boolean
   ): Promise<void> => {
+    if (this.state.editorComponentViewerDidAlreadyReload && !force) {
+      return;
+    }
     const component = viewer.component;
     this.application.componentManager.destroyComponentViewer(viewer);
-    this.setState({
-      editorComponentViewer: undefined,
-    });
-    this.setState({
-      editorComponentViewer: this.createComponentViewer(component),
-      editorStateDidLoad: true,
-    });
+    this.setState(
+      {
+        editorComponentViewer: undefined,
+        editorComponentViewerDidAlreadyReload: true,
+      },
+      () => {
+        this.setState({
+          editorComponentViewer: this.createComponentViewer(component),
+          editorStateDidLoad: true,
+        });
+      }
+    );
   };
 
   /**
@@ -972,11 +977,7 @@ export class NoteView extends PureComponent<Props, State> {
   render() {
     if (this.state.showProtectedWarning) {
       return (
-        <div
-          id="editor-column"
-          aria-label="Note"
-          className="section editor sn-component"
-        >
+        <div aria-label="Note" className="section editor sn-component">
           {this.state.showProtectedWarning && (
             <div className="h-full flex justify-center items-center">
               <ProtectedNoteOverlay
@@ -991,11 +992,7 @@ export class NoteView extends PureComponent<Props, State> {
     }
 
     return (
-      <div
-        id="editor-column"
-        aria-label="Note"
-        className="section editor sn-component"
-      >
+      <div aria-label="Note" className="section editor sn-component">
         <div className="flex-grow flex flex-col">
           <div className="sn-component">
             {this.state.noteLocked && (
@@ -1288,7 +1285,7 @@ export class NoteView extends PureComponent<Props, State> {
                   return (
                     <div className="component-view component-stack-item">
                       <ComponentView
-                        key={viewer.componentUuid}
+                        key={viewer.identifier}
                         componentViewer={viewer}
                         manualDealloc={true}
                         application={this.application}
